Moved gitfs code to a seperate module
